only norm left

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: ggrzesiek <ggrzesiek@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/08 11:33:25 by gkryszcz          #+#    #+#             */
-/*   Updated: 2025/03/10 07:35:35 by ggrzesiek        ###   ########.fr       */
+/*   Updated: 2025/03/10 13:33:34 by ggrzesiek        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,18 +16,19 @@
 # include <stdarg.h>
 # include <stdlib.h>
 # include <unistd.h>
+# include <stdint.h>
 
 int		ft_printf(const char *format, ...);
 void	ft_putchar_fd(char c, int fd);
 size_t	ft_strlen(const char *s);
-void ft_putnbr_fd(int n, int fd);
+void	ft_putnbr_fd(int n, int fd);
 void	ft_putstr_fd(char *s, int fd);
-size_t		num_len(long int n, int base);
+size_t	num_len(long int n, int base);
 char	*ft_itoa(int n);
 void	*ft_calloc(size_t nmemb, size_t size);
 void	ft_bzero(void *s, size_t n);
 int		sign_check(long int *tmp);
-int     ft_putnbr_hex_fd(unsigned n,int sign, int fd);
+int		ft_putnbr_hex_fd(unsigned long n, int sign, int fd);
 void	ft_putnbr_uns_fd(unsigned int n, int fd);
 
 #endif // FT_PRINTF_H
